requirements.txt
google-genai
langchain
pydantic
sqlalchemy
python-dotenv
pytest